requirements.txt
streamlit
diffusers
torch
transformers
pymongo
python-dotenv
bcrypt
Pillow
huggingface_hub


